move tiprackWellNamesByCol into fixtures/data.js
