+++ protocol-designer/src/step-generation/index.js
@@ -16,5 +16,4 @@ export {
 
 export * from './robotStateSelectors'
 export * from './types'
-export * from './data'
 export * from './utils'
